process_files_cloud_function/src/requirements.txt
google-cloud-contentwarehouse == 0.4.0
google-cloud-documentai == 2.20.0
google-cloud-storage == 2.7.0
google-cloud-logging == 3.5.0

